Move post_generated to a new decorators module
