+++ docs/reference/fields.rst
@@ -3,3 +3,6 @@ fields
 
 .. automodule:: polyfactory.fields
     :members:
+
+.. automodule:: polyfactory.decorators
+    :members:
