+++ polyfactory/__init__.py
@@ -1,6 +1,6 @@
 from .exceptions import ConfigurationException
 from .factories import BaseFactory
-from .fields import Fixture, Ignore, PostGenerated, Require, Use, post_generated
+from .fields import Fixture, Ignore, PostGenerated, Require, Use
 from .persistence import AsyncPersistenceProtocol, SyncPersistenceProtocol
 
 __all__ = (
@@ -10,7 +10,6 @@ __all__ = (
     "Fixture",
     "Ignore",
     "PostGenerated",
-    "post_generated",
     "Require",
     "SyncPersistenceProtocol",
     "Use",
